Refactored distance stuff
