+++ math/src/main/scala/breeze/linalg/package.scala
@@ -37,6 +37,21 @@ import breeze.linalg.DenseMatrix.canMapValues
  */
 package object linalg {
 
+  /*
+   * Distance metric types to use with generic distance function
+   */
+  type euclidean = euclideanDistance.type
+  type manhattan = manhattanDistance.type
+  type squared = squaredDistance.type
+  type tanimoto = tanimotoDistance.type
+  type minkowski = minkowskiDistance.type
+
+//  object euclidean extends DistanceMeasure
+//  object manhattan extends DistanceMeasure
+//  object squared extends DistanceMeasure
+//  object tanimoto extends DistanceMeasure
+//  object minkowski extends DistanceMeasure
+
   /**
    * Computes y += x * a, possibly doing less work than actually doing that operation
    */
